de-duplicate code to make codeclimate happy ... fixes https://github.com/dwyl/hapi-error/issues/12
